making code more DRY by making classes abstract
