+++ src/account_factory.py
@@ -1,4 +1,4 @@
-from src.interfaces import IAccount
+from src.abstract_account import AbstractAccount
 from src.account_mt5 import AccountMT5
 from src.account_sim import AccountSimulator
 
@@ -9,7 +9,7 @@ class AccountFactory():
     def __init__(self, production: bool):
         self.production = production
 
-    def create_account(self):
+    def create_account(self) -> AbstractAccount:
         if (self.production):
             return AccountMT5()
         else:
