+++ src/context_factory.py
@@ -1,5 +1,5 @@
 
-from src.interfaces import IContext
+from src.abstract_context import AbstractContext
 from src.context_mt5 import ContextMT5
 from src.context_sim import ContextSimulator
 
@@ -10,7 +10,7 @@ class ContextFactory():
     def __init__(self, production: bool):
         self.production = production
 
-    def create_context(self, json_settings: dict, credentials: dict) -> IContext:
+    def create_context(self, json_settings: dict, credentials: dict) -> AbstractContext:
         if (self.production):
             return ContextMT5(json_settings, credentials)
         else:
