+++ src/trade_execution_factory.py
@@ -1,4 +1,4 @@
-from src.interfaces import ITradeExecutor
+from src.abstract_trade_execution import AbstractTradeExecution
 from src.trade_execution_mt5 import TradeExecutorMT5
 from src.trade_execution_sim import TradeExecutorSimulator
 
@@ -9,7 +9,7 @@ class TradeExecutionFactory():
     def __init__(self, production: bool):
         self.production = production
 
-    def create_trade_executor(self, account: object):
+    def create_trade_executor(self, account: object) -> AbstractTradeExecution:
         if (self.production):
             return TradeExecutorMT5(account)
         else:
